requirements.txt
Django==2.1.4
django-pyodbc
django-pyodbc-azure==2.1.0.0
wagtail==2.4
python-crontab
wagtailmedia
django-cors-headers==2.4.0
django-storages==1.7.1
azure==4.0.0
coverage==4.5.1
gunicorn==19.9.0
flake8==3.6.0
django-debug-toolbar==1.10.1
factory-boy==2.11.1
django_extensions==2.1.3